Configure build step with Grunt
To build locally:

    npm install
    grunt

The compiled library is written to dist/DOMdiff-version.min.js. A
"require('DOMdiff')" is inserted at the end of the compiled file so
all of its modules are available by default.

The "dist" directory is not included in this pull request, but it
should be if everything checks out so anyone interested in DOMdiff
can grab the distribution version.

Fixes #30