Improved diff size by passing context data to roughlyEqual()
Adds a technique derived from the Patience Diff algorithm,
http://bramcohen.livejournal.com/73318.html. When comparing
subtrees, look for unique elements and consider them the same,
even if they have a different number of children. This is
implemented by generating valid CSS-like selectors for each tag
and discarding any that occur exclusively in one tree, or that
occur more than once in any tree.

The effect this has is that if we have a single div in both trees,
but one 3 children and one has 5, we don't want to remove one and
add the other, but instead diff the contents of each.

Similarly, if we have an input common to both trees that has an id
set, we can be sure that we'd rather add a p tag before it rather
than remove the input and create a new one that looks the same.
This is especially true for inputs since removing and re-adding a
form element can cause the element to lose focus.

A new test page was added that shows the number of affected
elements and rough diff size. Running this test with 1.0.0 and
this version should show a generally big improvement for the
example divs.

For situations where the current subtree has the same number of
elements with the same tag names and classes/ids, we also consider
such elements for child diffing instead of wholesale remove/add.